socket bind is not needed, plus fix timeouting on statusupdate
UDP is stateles, no need to try to bind - it is intended for assigning the socket to a specific interface on your machine.
Plus, if an update packet never arrives we do not want to remain stuck in this state.
(Git did not want o be nice about spliting it in two commits)